requirements.txt
pymongo
python-dotenv
pycodestyle
requests
jsonschema
wrapt
pytz
path
Flask
pycandela
IPython
Candela